requirements.txt
# Core
pygame-ce>=2.5.0

numpy>=1.24.0
pandas>=2.0.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Utils
python-dotenv>=1.0.0
requests>=2.31.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0
